Meetup group added to header
owasp-manama
The Meetup Pro account has an API that when the metadata is included in the index file will mirror the meeting information from the Meetup Group to the OWASP website.  If the meetup is used the metadata is not removed the leaders do not have to manually update.  See chapter policy for details.